+++ packages/mint-components/src/components/sqm-leaderboard/sqm-leaderboard.tsx
@@ -16,6 +16,14 @@ import { LeaderboardProps, useLeaderboard } from "./useLeaderboard";
 })
 export class Leaderboard {
   /**
+   * Filters leaderboard to show only data from this program. Defaults to the program context where this leaderboard is.
+   * If no program ID is set or provided by context, then a global leaderboard is shown.
+   *
+   * @uiName Program
+   */
+  @Prop() programId?: string;
+
+  /**
    * @uiName User Column Heading
    */
   @Prop() usersheading: string;
@@ -24,7 +32,7 @@ export class Leaderboard {
    */
   @Prop() statsheading: string;
   /**
-   * @uiName Rank Heading
+   * @uiName Rank Column Heading
    */
   @Prop() rankheading?: string;
   /**
@@ -33,7 +41,7 @@ export class Leaderboard {
   @Prop() showRank: boolean;
 
   /**
-   * Hide the current user's leaderboard information when they are not in the top 10
+   * Hide the viewer's leaderboard row if not in the top 10.
    *
    * @uiName Hide Viewing User
    */
@@ -46,13 +54,17 @@ export class Leaderboard {
    * @uiEnumNames ["Row Number", "Rank", "Dense Rank"]
    */
   @Prop() rankType: "rowNumber" | "rank" | "denseRank";
+
   /**
    * @uiName Leaderboard Type
    * @uiType string
-   * @uiEnum ["topStartedReferrers", "topConvertedReferrers"]
-   * @uiEnumNames ["Top Started Referrers", "Top Converted Referrers"]
+   * @uiEnum ["topStartedReferrers", "topConvertedReferrers", "topPointEarners"]
+   * @uiEnumNames ["Top Started Referrers", "Top Converted Referrers", "Top Point Earners"]
    */
-  @Prop() leaderboardType: "topStartedReferrers" | "topConvertedReferrers";
+  @Prop() leaderboardType:
+    | "topStartedReferrers"
+    | "topConvertedReferrers"
+    | "topPointEarners";
 
   /**
    * Title displayed for users without names
@@ -66,6 +78,13 @@ export class Leaderboard {
    * @uiWidgetOptions {"allowPastDates":true, "months": 1}
    */
   @Prop() interval: string;
+
+  /**
+   * Control the maximum amount of rows displayed on the leaderboard.
+   * 
+   * @uiName Maximum Rows Number
+   */
+  @Prop() maxRows: number = 10;
 
   /**
    * @undocumented
@@ -91,6 +110,8 @@ export class Leaderboard {
       rankheading: this.rankheading,
       rankType: this.rankType,
       leaderboardType: this.leaderboardType,
+      maxRows: this.maxRows,
+      programId: this.programId,
       anonymousUser: this.anonymousUser,
       interval: this.interval,
       hideViewer: this.hideViewer,
@@ -141,33 +162,33 @@ function useLeaderboardDemo(
     {
       firstName: "Viktor",
       lastInitial: "V",
-      value: 82,
+      textValue: "82",
       rank: 1,
       rowNumber: 1,
     },
     {
       firstName: "MF",
       lastInitial: "D",
-      value: 73,
+      textValue: "73",
       rank: 2,
       rowNumber: 2,
     },
     {
       firstName: "Freddie",
       lastInitial: "G",
-      value: 64,
+      textValue: "64",
       rank: 3,
       rowNumber: 3,
     },
     {
       firstName: "Benny",
       lastInitial: "B",
-      value: 55,
+      textValue: "55",
       rank: 4,
       rowNumber: 4,
     },
     {
-      value: 46,
+      textValue: "46",
       rank: 5,
       rowNumber: 5,
     },